refactor: merge monorepo-utils into plugin-source-build
