--- packages/monorepo-utils/modern.config.ts
@@ -1,3 +1,0 @@
-import { configForDualPackage } from '../../scripts/modern.base.config';
-
-export default configForDualPackage;
