--- packages/monorepo-utils/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "@rsbuild/monorepo-utils",
-  "version": "0.7.0-beta.3",
-  "description": "The monorepo utils of Rsbuild.",
-  "homepage": "https://rsbuild.dev",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/web-infra-dev/rsbuild",
-    "directory": "packages/monorepo-utils"
-  },
-  "license": "MIT",
-  "type": "module",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "import": "./dist/index.js",
-      "require": "./dist/index.cjs"
-    }
-  },
-  "main": "./dist/index.cjs",
-  "types": "./dist/index.d.ts",
-  "files": [
-    "dist"
-  ],
-  "scripts": {
-    "build": "modern build",
-    "dev": "modern build --watch"
-  },
-  "dependencies": {
-    "@rsbuild/shared": "workspace:*",
-    "fast-glob": "^3.3.2"
-  },
-  "devDependencies": {
-    "@types/node": "18.x",
-    "typescript": "^5.4.2"
-  },
-  "publishConfig": {
-    "access": "public",
-    "provenance": true,
-    "registry": "https://registry.npmjs.org/"
-  }
-}
