--- packages/monorepo-utils/src/index.ts
@@ -1,5 +1,0 @@
-export * from './common';
-export * from './project';
-export * from './project-utils';
-export * from './constants';
-export * from './types';
