--- packages/monorepo-utils/tsconfig.json
@@ -1,18 +1,0 @@
-{
-  "extends": "@rsbuild/tsconfig/base",
-  "compilerOptions": {
-    "outDir": "./dist",
-    "baseUrl": "./",
-    "rootDir": "src",
-    "composite": true
-  },
-  "references": [
-    {
-      "path": "../core"
-    },
-    {
-      "path": "../shared"
-    }
-  ],
-  "include": ["src"]
-}
