+++ packages/plugin-source-build/src/common/getBaseData.ts
