+++ packages/plugin-source-build/src/common/getProjects.ts
